requirements.txt
airtable-python-wrapper==0.12.0
google_auth_oauthlib==0.4.1
google-api-python-client==1.7.11
pandas==0.25.3
pygsheets==2.0.2
pytz==2019.3
pyyaml==5.2
requests==2.22.0
sqlalchemy==1.3.12
tabulate==0.8.6
xlsxwriter==1.2.7
